Update ./gradlew versions task
The versions task works mostly the same as it did before. The main difference is how we link a Neo4j version to an APOC Version. The logic here used to be a bit complex because of various edge cases, but going forward APOC will guarantee that it is backwards compatible with any known major version of APOC.