Fix versions (#167)
* Cleanup 3.x documentation

Docs for APOC 3.x don't seem to be hosted anywhere. In any case, even if they were hosted they would be hosted from neo4j-contrib/apoc. As such, these files can be deleted from neo4j/apoc.

* Update ./gradlew checksums task

The task works nearly the same as it did before. The only difference is updating URL references and adding TODO to delete the task once we begin signing jars.

* Update ./gradlew versions task

The versions task works mostly the same as it did before. The main difference is how we link a Neo4j version to an APOC Version. The logic here used to be a bit complex because of various edge cases, but going forward APOC will guarantee that it is backwards compatible with any known major version of APOC.

* Commit updated checksums.json and versions.json

* Normalize version tags

Removed some unecessary normalization of "beta" tags

* Ensure versions.json didn't shrink

We had some cases in the past where the versions file was wiped by TeamCity. We added a check to make sure that can't happen.

* Minor bug fixes