[UHxIZ8JO] Ignore PRs from JLLeitschuh/security-research
We want to ignore PRs from this security researcher because they do not follow our private disclosure process. These PRs highlight that our code is exploitable before we've had the opportunity to fix it.

However, we will take JLLeitschuh up on his suggestion to add https://github.com/github/codeql-action to run checks in the background.