[UHxIZ8JO] Add CodeQL (#186)
* [UHxIZ8JO] Ignore PRs from JLLeitschuh/security-research

We want to ignore PRs from this security researcher because they do not follow our private disclosure process. These PRs highlight that our code is exploitable before we've had the opportunity to fix it.

However, we will take JLLeitschuh up on his suggestion to add https://github.com/github/codeql-action to run checks in the background.

* [UHxIZ8JO] Install CodeQL

* [UHxIZ8JO] Remove unused file

The file was removed because it contained some vulnerabilities, and also because it was not used. It's also not used in APOC Extended.