[f3ZUxEG2] Fix bugs and refactor apoc.cypher.runMany
Fixes two issues in apoc.cypher.runMany:

- Transactions were accessed from multiple threads,
  this is not allowed and has undefined behaviour.
- Errors were quietly ignored.
  This was not documented so I assume it was an oversight.

For simplicity, the implementation is changed to be single threaded.
We previously spawned a second thread for executing inner queries.
While this could have performance benefits in some situations,
there is also an overhead of starting and synchronizing threads.
We currently have no benchmarks for this procedure so there's not
much to base this decision on either way.

Some undocumented behaviour is kept from previous implementation:

- Ignore schema changing queries.
- Run statements in new transactions.

Refactored some tests to always display the full result set on failure.